Added check if file exists to nnunet preprocessing

+++ breast_fgt_segmentation/preprocess_for_nnunet_inference.py
@@ -6,7 +6,8 @@ import os
 def main(input_folder: str, output_folder: str):
     for patient_folder in os.listdir(input_folder):
         first_timepoint_path = os.path.join(input_folder, patient_folder, f'{patient_folder.lower()}_0000.nii.gz')
-        shutil.copy2(first_timepoint_path, output_folder)
+        if os.path.exists(first_timepoint_path):
+            shutil.copy2(first_timepoint_path, output_folder)
 
 
 if __name__ == '__main__':
